modify user adding sex feature
