+++ capstone/src/main/java/com/capstone/forum/treeholeforum/CapstoneApplication.java
@@ -1,0 +1,13 @@
+package com.capstone.forum.treeholeforum;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CapstoneApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CapstoneApplication.class, args);
+	}
+
+}
