+++ capstone/src/main/java/com/capstone/forum/treeholeforum/DTO/MomentBody.java
@@ -1,0 +1,11 @@
+package com.capstone.forum.treeholeforum.DTO;
+
+import java.util.List;
+import lombok.Data;
+
+
+@Data
+public class MomentBody {
+    private String momentName;
+    private List<String> tags;
+}
