+++ capstone/src/main/java/com/capstone/forum/treeholeforum/DTO/RegReqBody.java
@@ -1,0 +1,12 @@
+package com.capstone.forum.treeholeforum.DTO;
+
+import lombok.*;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class RegReqBody {
+    private String username;
+    private String email;
+    private String password;
+}
