+++ capstone/src/main/java/com/capstone/forum/treeholeforum/configuration/PwdSecurityConfig.java
@@ -1,0 +1,27 @@
+package com.capstone.forum.treeholeforum.configuration;
+
+import org.springframework.context.annotation.*;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+//import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.argon2.Argon2PasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.config.annotation.web.configuration.*;
+
+public class PwdSecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Override
+    public void configure(HttpSecurity httpSecurity) throws Exception {
+        httpSecurity.csrf().disable()
+                .authorizeRequests()
+                .antMatchers("/api/auth/**")
+                .permitAll()
+                .anyRequest()
+                .authenticated();
+    }
+
+    @Bean
+    PasswordEncoder passwordEncoder() {
+        Argon2PasswordEncoder argon2PasswordEncoder = new Argon2PasswordEncoder();
+        return argon2PasswordEncoder;
+    }
+}
