+++ capstone/src/main/java/com/capstone/forum/treeholeforum/model/VerifyMail.java
@@ -1,0 +1,12 @@
+package com.capstone.forum.treeholeforum.model;
+
+import lombok.*;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class VerifyMail {
+    private String subject;
+    private String recipient;
+    private String body;
+}
