+++ capstone/src/main/java/com/capstone/forum/treeholeforum/model/VerifyToken.java
@@ -1,0 +1,26 @@
+package com.capstone.forum.treeholeforum.model;
+
+import java.time.*;
+
+import lombok.*;
+
+import javax.persistence.*;
+
+import static javax.persistence.FetchType.LAZY;
+import static javax.persistence.GenerationType.IDENTITY;
+
+
+@Data
+@Entity
+@AllArgsConstructor
+@NoArgsConstructor
+@Table(name = "token")
+public class VerifyToken {
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+    private String token;
+    @OneToOne(fetch = LAZY)
+    private User user;
+    private Instant expiryTime;
+}
