+++ capstone/src/main/java/com/capstone/forum/treeholeforum/model/VoteType.java
@@ -1,0 +1,9 @@
+package com.capstone.forum.treeholeforum.model;
+
+public enum VoteType {
+    UPVOTE(1), DOWNVOTE(-1),
+    ;
+
+    VoteType(int type) {
+    }
+}
