+++ capstone/src/main/java/com/capstone/forum/treeholeforum/repo/UserRepository.java
@@ -1,0 +1,10 @@
+package com.capstone.forum.treeholeforum.repo;
+
+import com.capstone.forum.treeholeforum.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+}
