--- capstone/src/main/java/com/example/capstone/CapstoneApplication.java
@@ -1,13 +1,0 @@
-package com.example.capstone;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CapstoneApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CapstoneApplication.class, args);
-	}
-
-}
