--- capstone/src/main/java/com/example/capstone/DTO/MomentBody.java
@@ -1,11 +1,0 @@
-package com.example.capstone.DTO;
-
-import java.util.List;
-import lombok.Data;
-
-
-@Data
-public class MomentBody {
-    private String momentName;
-    private List<String> tags;
-}
