--- capstone/src/main/java/com/example/capstone/DTO/RegReqBody.java
@@ -1,12 +1,0 @@
-package com.example.capstone.DTO;
-
-import lombok.*;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegReqBody {
-    private String username;
-    private String email;
-    private String password;
-}
