--- capstone/src/main/java/com/example/capstone/model/VerifyMail.java
@@ -1,12 +1,0 @@
-package com.example.capstone.model;
-
-import lombok.*;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class VerifyMail {
-    private String subject;
-    private String recipient;
-    private String body;
-}
