--- capstone/src/main/java/com/example/capstone/model/VerifyToken.java
@@ -1,31 +1,0 @@
-package com.example.capstone.model;
-
-import java.time.*;
-
-import lombok.*;
-import org.springframework.lang.Nullable;
-
-import javax.persistence.*;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotEmpty;
-import java.util.*;
-import static javax.persistence.FetchType.LAZY;
-import static javax.persistence.GenerationType.IDENTITY;
-import static javax.persistence.GenerationType.SEQUENCE;
-
-
-@Data
-@Entity
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "token")
-public class VerifyToken {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-    private String token;
-    @OneToOne(fetch = LAZY)
-    private User user;
-    private Instant expiryTime;
-}
