--- capstone/src/main/java/com/example/capstone/model/VoteType.java
@@ -1,9 +1,0 @@
-package com.example.capstone.model;
-
-public enum VoteType {
-    UPVOTE(1), DOWNVOTE(-1),
-    ;
-
-    VoteType(int type) {
-    }
-}
