--- capstone/src/main/java/com/example/capstone/repo/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.capstone.repo;
-
-import com.example.capstone.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String username);
-}
