--- capstone/src/main/java/com/example/capstone/repo/VoteRepository.java
@@ -1,9 +1,0 @@
-package com.example.capstone.repo;
-import com.example.capstone.model.*;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.*;
-
-public interface VoteRepository extends JpaRepository<Vote, Long>{
-    Optional<Vote> findTopByPostAndUserOrderByVoteIdDesc (Moment moment, User user);
-}
