+++ capstone/src/test/java/com/capstone/forum/treeholeforum/CapstoneApplicationTests.java
@@ -1,0 +1,13 @@
+package com.capstone.forum.treeholeforum;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CapstoneApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
