--- capstone/src/test/java/com/example/capstone/CapstoneApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.capstone;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CapstoneApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
